Rename `SsoHookHandler` to `SharedDomainHookHandler`
Just like `SharedDomainUtils`, this patch renames the shared domain
hook handler to something more descriptive and not attached to the
previous SSO acronym we use to have before.

The name of the shared domain might not have SSO in it so we should
no longer push this idea in the code.

Change-Id: I8847c1f935d58d5a72fb684aadd53f17b8d2b756