Rename CentralAuthSsoPreAuthenticationProvider class
Follow-up: I8847c1f935d58d5a72f, in the spirit of removing the SSO
acronym from CA related to SUL3, renaming this class to replace the
acronym patch to "SharedDomain" which is more descriptive to what we
want.

Also, updated tests that use `sso.` which really is not a big deal
because they're just test but well, let's just replace them.

Dealing with config settings will happen in a different patch or
different series of patches as migrating that is not as easy as just
doing a rename.

Change-Id: Ic836b878295db38445ada2577a23e461ed1c6466