Rename MWCU.php -> MWCA.php
MWCU was created in Icdda4a26447b0570b778 as a utility shortcut for
typing long class names including namespaces when debugging in places
like Shell.php or Eval.php.

But Taavi pointed out that CU could be interpreted as CheckUser because
we do have an extension like that and maybe other things in our context.

CA is already widely understood in our context as CentralAuth and
hence renaming the class to `MWCA` which also reads nicely when we do
`MWCA::user()` - MediaWiki CentralAuth user.

Change-Id: I27d6843b172baa621d1b0446145e0e81d49f01a9