Set 'remember' central session object field when recreating
When the central session expires but there is a centralauth_Token
cookie, the central session gets automatically recreated, but the
'remember' field in the central session data gets lost. This can
lead to autologin setting up non-remembered sessions on other
domains, even though the user asked to be remembered.

Avoid this by copying the local session's remembered flag
(set based on the presence of the centralauth_Token cookie) into
the central session data.

Test plan: get a logged-in user, delete all session cookies on
local wiki, delete centralauth_Session and <wiki>_session
(but not centralauth_Token etc) on central login wiki. Wait for
the autologin, verify that local wiki ends up with a _Token cookie.

Bug: T379254
Bug: T372702
Change-Id: If1bfe7ebaa637b15985109ecb69505cead6fc644