Use cookie to access central session when local session expired
Normally the central session ID is stored in the local session,
and the central session manager relies on that. But it's also
stored in a cookie, and that cookie is more durable than the
local session (the local and central session cookies both have
session expiry, but the local session object expires after one
hour of inactivity) so fall back to the cookie when the session
is not found.

This is important because, when the local session expires, it
will be re-persisted, and we don't want the central session ID
to change when that happens.

Test plan: get a logged-in user, delete <wiki>_session cookie
(but not the centralauth_* cookies), reload, verify that
the centralauth_Session cookie did not change.

Change-Id: Ia1b205736ad3cf0eb924fd8f413fabf18436cad4
(cherry picked from commit 117c514de4e57098e404d04ed8f6f854f7810f64)