SUL3: Unbreak action API requests in SUL3 mode
The action API should behave in SUL3 as if we're in SUL2 mode, so
that action API end-points such as `clientlogin` and the rest of it
doesn't break.

Basically, this patch should make SUL3 behave like a no-op/SUL2 even
if SUL3 is enabled.

This functionality got broken because in SUL3 mode, a button authentication
request got returned in the case of `clientlogin` instead of password
authentication request. This patch unbreaks API request in SUL3.

Test plan
=========

0. Make sure SUL3 mode is disabled via `$wgCentralAuthEnableSul3`.
1. Visit Special:ApiSandbox of a local wiki.
2. Utilize the `clientlogin` API to log a user into the wiki.
3. Verify that the user is logged-in using username and password.
4. Enable SUL3 mode and retry step 1 - 3 above.
5. Verify that the user is logged-in the same way as step 3 above.

Bug: T379816
Change-Id: Ia1c3183bdd68497b2358116c785d3ca0c1dee6fe