maintenance: Cleanup usage of deprecated `::getSelectQueryData()`
... method from DatabaseLogEntry class. There is a recommended one
to migrate to in LoggingSelectQueryBuilder.

NOTE: `$queryData['options']` previously returned an empty array,
so we can ignore that in the new query. Join conditions are already
taken care of by the logging SQB, amongst other things.

Test plan
=========

The script should still work as before, this should be a no-op
in terms of impact.

Change-Id: I54c1f4fa6b88db523326ae02573e444970fbf0a3