SUL3: Don't fallback to SUL2 login wiki for temp accounts
Since we're not accepting SUL3 mode for API clients and fallback
to SUL2, there is a special case where editing with VisualEditor
happens async HTTP request via the api.php entry point for temp
accounts. We don't want to pretend as if we're not in SUL2 mode
here because the edit is not happening on the shared domain but
only comes from the API endpoint.

We want to accept it and detect the SUL mode since temp accounts
are created by Special:CentralLogin on the central wiki and we will
like to use the correct shared domain in this case rather than just
always default to SUL2 mode.

Before this patch, making an edit with a temporary account for the
first time for example and CA trying to perform central login, the
SUL2 wiki is always used but with this patch applied, the correct
shared domain is used based on the SUL mode (shared domain for SUL3
and central login wiki for SUL2).

Also, with this patch, it seems like edge-login for other autologin
wikis also work correctly as now we're able to correct always use
the correct auth domain for cookies refresh etc.

Bug: T383812
Change-Id: Ibe07263e2460daafc686421b06da5a86315be095