Do not append query params multiple times during URL generation
Found while going through logstash that some URLs are not as they
should be.

`SpecialCentralAutoLogin::do302Redirect()` will try to append the
URL query parameters a second when about to perform the redirect.
This is no longer needed as `CentralDomainUtils::getUrl()` already
does the correct thing internally.

Doing it a second time is harmless but we don't want to mess with
links like this. Maybe it even has some side effects.

Without this patch, we have URL endpoints constructed such as:
`/setCookies?from=enwiki&from=enwiki&type=1x1&type=1x1` but what
we really want is: `/setCookies?from=enwiki&type=1x1`.

This patch fixes the URL generation logic to do the correct thing.

Change-Id: Ib2ea65eac9241ceaee8c674b411cfed431724e0d