SUL3: Allow temp users to authenticate (login/signup) via the API
When a temp user has an active session, trying to authenticate with
this fails due to Ibe07263e2460daafc686421b06da5a86315be095. We need
to fallback to SUL2 if the user is a temp account that wants to gain
access to their permanent/named user account.

Test strategy
=============

0. Enable SUL3.
1. Edit any page you have locally as an anon user to gain a temp session.
2. Visit the /wiki/Special:ApiSandbox and attempt a `clientlogin` action.
3. You should be able to login with the `clientlogin` API.

Without this patch, doing the above steps fail because the code tries to
run in SUL3 mode instead of falling back to SUL2 for temp accounts.

Bug: T384523
Change-Id: I7fb0029078e1706b885f415f62e81906f4c86ec5