Add configuration options and global preference for the SUL3 rolllout
This patch adds new config vars "Sul3RolloutUserPercentage",
"Sul3RolloutAnonSignupPercentage" and "Sul3RolloutSignupCookie"
to help manage the SUL3 rollout, a new global preference
"centralauth-use-sul3" to be set if a user signing up or logging in
is designated to be in the rollout, and a temporary "sul3wanted" cookie
to assist in the IP user signup case.

None of this is bullet-proof; there wil be plenty of edge cases where
a user on one device or having first visited a non-SUL3 enabled wiki
may get an inconsistent experience. But trying to make this bullet-proof
would make things even harder to reason about than they already are.

Bug: T384549
Bug: T377144
Bug: T384552
Bug: T384215
Depends-On: I3ead8cc1371352896ac5458ab85c31d0f3640c2c
Change-Id: I241872237a8b27bc4261e6231b55b068bd3ac428