Do not preserve 'sul3-action' when restarting authentication
'sul3-action' tells whether the action was login or signup; the
reset=true option of AuthPreserveQueryParams means that authentication
is restarted, and possibly switched to the other action, so this
parameter shouldn't be preserved.

This doesn't have any functional effect because the authentication
pages redirect immediately, so there is no way to switch between them.
But it's conceptually more correct.

Follows up I3709c987774c79048ed58f870278a78db7039ec0.

Bug: T364866
Change-Id: Ib71caa2b458a397d2c5a3f9c989b466e25818f62