Add passive central domain to edge login list
When doing edge login, do an edge login on the "wrong" central
domain (the SUL2 domain if we are in SUL3 mode, the SUL3 domain
if we are in SUL2 mode). This helps keeping the two domains in
sync.

Also clean up CentralDomainUtils tests so it's clearer what they
are testing.

Bug: T375796
Change-Id: Iddc22e2dcfc2a2af213b24133d5937491f716968