Make sure isSul3Enabled() is a boolean
I24187223 changed SharedDomainUtils::isSul3Enabled() to tri-state
(true/false/null) but that broke some usages which passed the
return value to EventLogging, which has strict types.

Because the affected callers are in several different repositories,
rather than fixing them, it's easier to restore the original return
type of isSul3Enabled() and use an optional output variable for the
third state.

This also fixes a bug where, because of a non-strict type comparison
at the beginning of onSpecialPageBeforeExecute(), an explicit
usesul3=0 URL flag was ignored in some cases.

Also simplify the logic in some places, and make it possible to use
the 'sul3OptIn' cookie for opting out as well as opting in.

Bug: T384549
Change-Id: Icd23fc7d9e616b5de15d9d4cfe95140fcf22968c