CentralAuthIdLookup: Use primary DB after writes
Following the lead of CentralAuthUser, use the primary DB for
CentralIdLookup reads when there were recent writes to the central
DB. (Or more precisely the load balancer the central DB is on;
this is somewhat imprecise when the central DB shares an LB with
wiki databases.)

This was already the case for most CentralAuthUser-related
queries, but not 1) queries in CentralAuthIdLookup itself;
2) for deciding whether to use the WAN cache; 3) for deciding
whether objects in the in-process cache can be reused.

This removes most of the difference between centralIdFromName()
batch and non-batch queries, and between centralIdFromName() and
nameFromCentralId(), in terms of how much they are affected by
replication lag.

Probably this logic should be more centralized, but this is a start.

Bug: T379909
Bug: T380500
Change-Id: I1f9a6b518aa74798fa7c8e2172e6c8dfdf7d06bf