Move the CentralAuthUser instance cache to a service
For lifetime management in tests, it is better to use a service than a
static member variable.

Of course there was a test depending on the user cache surviving a
service reset. Fix that.

Increase the cache size from 20 to 100, since GenderCache is going to be
hitting it pretty hard.

Change-Id: Iab263fe7495e3381605ba53a4fb62e2c9e3747ad