tests: Clarify array shapes of parameters in a helper method
$methods is a list (numeric array), and array `+` operator will omit
keys that are already in the array, including numeric keys, resulting
in some of the "basic" methods defined in this test being disallowed.
We want to concatenate the arrays instead with `...` (or array_merge).
There is only one test using this parameter, and it was not affected
by the bug.

$return is an associative array, so clarify that too.

Change-Id: I397900ee3510b73378c6891ac73b9771e4ff94d0