SUL3: Fix default for `wgCentralAuthLoginWiki`
Code consuming `wgCentralAuthLoginWiki` has a fallback to the current
domain or the domain we're coming from using the `from` query parameter
on the request object.

This has been so since in SUL3 we have multiple shared/central domains
at play, with one active and another passive. But in the case where we
do not have a central domain configured, we still fallback to the current
domain to be used for edge-logins.

Bug: T388218
Change-Id: Iefe1cbd6f9ec665a03fa6fc2ce7fba924310c99a