Try both SUL2 and SUL3 central domain for autologin
When the /checkLoggedIn step for central autologin fails, try
it on the passive central domain, while making sure anonymous
responses remain cacheable:
* Introduce two more CENTRAL_DOMAIN_ID constants for
  deterministically referencing the SUL2 and SUL3 domains.
* Make sure all references to /start or /checkLoggedIn include an
  explicit 'usesul3' query flag.
* Make sure the redirect target of /start is fully determined by
  the URL. This fixes an error we introduced earlier, since these
  redirects usually get cached (although it's somewhat superfluous
  given the change to the references).
* Make /checkLoggedIn on the SUL2 domain redirect to /checkLoggedIn
  on the SUL3 domain if the first attempt doesn't yield a user, and
  vice versa. Use an additional query flag to prevent loops.

This change will affect all users regardless of SUL3 opt-in status
(not really avoidable since when autologin starts we don't know
who the user is), so we'll need to be careful when deploying it.
Also, as the /start and /checkLoggedIn URLs get cached, these
code changes don't necessarily take immediate effect, or don't
necessarily take effect at the same time for different users or
domains or steps. I think the change is safe, even given that.

Bug: T375796
Change-Id: I545031e888e9081a5f0165223040cf9b7e1388cf