Do not trigger edge login on the shared domain
Prevent edge login when the user is on the shared domain.
It caused errors in Wikimedia setup due to some config pecularities
(although that will be fixed separately), and also is just pointless
in general (it would just error out due to the assertIsLocalDomain()
checks).

One pathway was setting a session flag in postAuthentication() in
CentralAuthPrimaryAuthenticationProvider (note that that will be
called regardless of which primary provider handled the
authentication, so it will still set the flag during the local leg
of a SUL3 login). The other pathway was an edge login to the
passive central domain.
(Not sure the whole idea of scheduling another edge login at the
end of an edge login makes sense, but leaving that for another
time.)

Change-Id: I96dedc53468fd6f75ab843591e445e3576059972
(cherry picked from commit 8b7bb772f9badf492736af60817e5db95427e98f)