Do not initiate central login on the passive central domain
This is mostly for the benefit of Wikimedia production where
we have some automated tests that don't expect that to happen.
But in general, it doesn't make a lot of sense and makes the
system even more complicated.

Also disable JS autologin on both central domains. JS is cached
and does not depend on the user, so "active" and "passive" are
not meaningful concepts in a ResourceLoader callback.

Follows up I96dedc53468fd6f75ab843591e445e3576059972 which missed
these cases.

Bug: T388218
Change-Id: I80f3fcc07c5608e7c447af78f70889ec809edbda