Do not throw an exception after shared-domain login with no token
...just assume this is not part of a local login flow and the user
was trying to log in specifically on the shared domain (e.g. as
part of a credentials change).

This probably obsoletes some (all?) of the other conditions for
returning early from the hook, but trying to keep the change
minimal for now since it's needed as a hot fix for accessing 2FA
management pages on the central domain.

Bug: T362715
Change-Id: I9b9318cea9f2f40032cf18a7ae72b643f0280ed4