Do not start central login from the shared domain
This wasn't happening during SUL3 login (because isSul3Enabled() is
always true then) but will happen when someone tries to log in on
the central domain directly, e.g. for credentials change, and will
end up with an "No active login attempt is in progress for your
session"error message (shown on the local domain).

Bug: T362715
Change-Id: Id39f647942c4b9f278852260d20486fae4e790d2