Remove redundant WaitConditionLoop from CentralAuthTokenManager
Similar to I97d8f108de (52af356cad) in MediaWiki core. This existed
to support an original 2015 draft for Multi-DC. At the time, there was
not yet a "MicroStash" and there was not yet a "TokenManager". Instead,
tokens were stored in the MediaWiki core session store (e.g. Cassandra
at WMF). And like other session store writes, these go to the local
DC, and get eventually replicated to other DCs, so for reliable reads
one would need to either use READ_LATEST or wait for replication.

In 2022, as Multi-DC turned from theory to reality, this plan changed
in T278392 where we decided to introduce a separate token store
concept for various cases (including CentralAuth) by re-using the
existing "mcrouter-primary-dc" route to Memcached. This was later
recognised in MediaWiki core under the name "MicroStash", which,
unlike the SessionStore, is meant to be simple, centralised, and
without any observable replication complexity for the consumer.

== History ==

* 2019 (I4e379244da, 15be87509b):
  CentralAuth stored tokens reside in the MW core SessionStore
  (Cassandra at WMF), for which MediaWiki supports both dc-local
  (READ_NORMAL) and dc-primary (READ_LATEST).

  This change switched from READ_LATEST, which would in theory do a
  cross-dc read, to instead do explicit local reads. Note that this
  was a no-op change because in 2019 we were not yet Multi-DC so
  READ_LATEST is the same as a local read since the only active DC
  was the primary DC.

  On the assumption we'd use a replicated session store for tokens,
  this patch also added back use of WaitConditionLoop, to do retries
  in case the key isn't there yet (i.e. if we create the token in
  one DC, and then read it in another before it finishes replication).

* July 2022 (T278392, Ia4fefb0eef0ab and I7dbb5e33a1):
  Change $wgCentralAuthTokenCacheType in wmf-config from Redis
  to mcrouter-primary-dc. This was the final Multi-DC plan,
  which does not involve replication for micro stashing. At this
  point, we could have removed the WaitConditionLoop.

* Jan 2024 (Id67dd87fe3b1):
  Switch from "CentralAuthTokenStore" (defaulting to SessionStore,
  but overridden at WMF to use mcrouter-primary-dc) to MediaWiki core
  MicroStash (which is also mcrouter-primary-dc at WMF).

  While a no-op for WMF, this change removed the last remaining
  justification to wait for replication (which you can reasonably
  expect from a SessionStore fallback at third parties, and from
  Cassandra at WMF). MicroStash instead promises a central interface
  without consumer-side replica complexity.

* Aug 2024 (I560551f49a, e59ff145fb):
  In CR there, I suggested we remove this code, which I'm now doing.

Change-Id: Ia1a31b1c4516115f116ed4d9852f3bc940d71216