SUL3: Fix user ID mismatch during login (immediately after creation)
Immediately after account creation, MW will try to log the user in by
visiting `Special:UserLogin` in the `continuePrimaryAuthentication()`
call but that can be affected by replication lags which will cause us
seeing the central user ID as `0` (doesn't exist yet) because calling
`getInstanceByName()` on CentralAuthUser will result to using a replica
DB.

We want to use a primary DB instead since we just created a user and
must see it immediately when authenticating them on the local domain.
Let's see if this is going to resolve the issue.

Normally, ChronologyProtector is supposed to protect us againsts these
kinds of issues but at the moment, it doesn't forward related cookies
upon redirect from the shared domain to the local domain where the auth
process finishes. This may need fixing at some point.

NOTE: For a regular login (not preceded by a signup), using the replica
DB is fine. We only want to use the primary DB immediately after a signup.

Bug: T388177
Change-Id: I826a4a113f660255c8cb450a22a34a54762ea8e8