SUL3: Retry local login on failure due to invalid/expired login token
Before this patch, we let the user know that their centralauthLoginToken
is invalid or has expired but we can do better.

The scenario above could happen when the login flow in SUL3 takes
some time to be completed, most likely during a MFA login flow
where the user could likely spend some time before they enter
their MFA key/code after successfully filling in their username/password
and get provided the MFA form. In between the password step and
theMFA step, MicroStash[1] which we use at WMF to store login
tokens (backed by memcached) can wipe out (due to it's LRU protocol)
the user's login token. So even after filling in the correct MFA
(in our WMF case - OATHAuth), the user gets logged into the shared
domain but continuing the authentication on the origin wiki fails
(due to the centralauthLoginToken being invalid).

If the login fails on the local domain, lets retry it in the
background by performing a GET redirect to the local `Special:UserLogin`
that will automatically do a central session fetch (with a
new token) and since there is already one that exists (at this
point in time), we'll just log the user into the origin wiki.

To test this, I did the following:
1. Installed the OATHAuth extension to enable MFA for a
   given account
2. Setup memcached or any other cache backend to use as
   some short-lived storage for token store. Something
   you can access and delete key/value pairs.
3. Enable MFA for a named account.
4. Clear all cookies and start a new login flow
5. Click on login on a local wiki
6. Fill in the username and password on the shared domain
7. Connect to a MW shell and delete the centralauthLoginToken
   for the request in (6) above. NOTE: Delete from the
   store you configured in (2) above.
8. Fill in the MFA key/token in the MFA view and click
   submit.
9. Step 8 above will succeed but the local leg of the login
   will fail since the token has been deleted then an automatic
   retry of the local login will trigger.
   Since there is already a central session at this point,
   the local login retry will be quick and will most likely
   succeed, then the user will be logged-in locally.

NOTE:
     Initially, I wanted to read the value of the wiki ID from
     the $wgCentralAuthSharedDomainCallback config setting, see
     https://gerrit.wikimedia.org/r/c/mediawiki/extensions/CentralAuth/+/1135511/12
     but that may be fragile depending on how sysadmins/devs
     configure their shared domain.

[1] https://www.mediawiki.org/wiki/Object_cache#MicroStash

Bug: T390784
Change-Id: I2af968a5a01c602525b1902f832ee1b7660e38ec