Simplify remaining CentralAuthEnableSul3 options
Remove support for 'query-flag' value and enable that logic
unconditionally. It was already documented that configuring SUL3
(i.e. CentralAuthSharedDomainCallback) without enabling that value
would cause the system to function incorrectly. Remove the redundancy,
and just check for whether the shared domain callback is configured.

That leaves [] and ['always'] as the only valid values for
CentralAuthEnableSul3, so simplify it and make it a boolean,
like it was when it was first introduced.

Bug: T355281
Change-Id: I1efa03c22c3c428004530fef79f23bca47c84add