Hide sitenotices on the central domain and show WebAuthn notice
Do not show normal site notices on the central domain. The
GetLocalURL hack for making links point back to the local domain
probably won't work for them, and in any case they are a distraction.

Show a programmatic site notice on the WebAuthn management page,
linking to the other version (central or local) of the page.
(It would be nice to only do this when the user has old WebAuthn
keys, and otherwise just redirect them to the central domain,
but OATHAuth/WebAuthn doesn't expose that information.)

Bug: T390035
Bug: T362715
Change-Id: I3b2732cfdb7a501a18b2ad2e3296e8d45cb16b17