Simplify authentication provider filtering
Rather than listing providers that are disallowed on the local domain
(which is basically all of them), only list ones which are allowed.
This way, we don't need to reference a bunch of non-CentralAuth
classes, and the list doesn't need to be updated every time something
new is deployed.

(It's still possible we'll want a non-CentralAuth provider in the
list some day, e.g. for logging, or for preventing abusive temp
user creations, but the $wgCentralAuthSul3SharedDomainRestrictions
mechanism should have no trouble handling that.)

In Wikimedia production, this filters some providers that weren't
filtered before:
* WikimediaEvents\AccountCreationEventProvider
* MediaWiki\Extension\Campaigns\CampaignsSecondaryAuthenticationProvider
* MediaWiki\Auth\ResetPasswordSecondaryAuthenticationProvider
* MediaWiki\Auth\EmailNotificationSecondaryAuthenticationProvider

This should be fine. ResetPasswordSecondaryAuthenticationProvider
only triggers when another provider sets a session flag (which none
of the non-filtered providers do), the others only implement callbacks
for account creation (which never happens locally in SUL3).

Security-wise, an allowlist is worse than a disallow-list, because the
failure mode is providers (that might need to block authentication)
unexpectedly missing. But we already ensure that that cannot happen
for CentralAuthSharedDomainPreAuthenticationProvider, and the other
allowed ones are not security-sensitive (although
CentralAuthPrimaryAuthenticationProvider disappearing would result in
some DB inconsistencies). And since all of the allowed providers are
within CentralAuth, it's much harder for that kind of error to
happen.

Bug: T390329
Change-Id: I7793a17fbbefedbdce1dc0902a8e4018fc0b03f8