SUL3: Fix account creation by username & email (with temp password)
There are 4 ways to create an account, see: https://www.mediawiki.org/wiki/Manual:User_creation
and "byemail" is one of them where the user creating the account must
be an authenticated user.

Whether via the API or regular web UI, the user needs to be logged-in
to use this feature and under SUL3 mode, using the API, we didn't take
care of this scenario for named accounts.

Before this patch, the exemption was applied to user types but that
doesn't scale well and creates some maintainability burden. Let's just
handle the exemption based on the request type (API).

To test this, visit the API sandbox and construct a request with the
following parameters:

`action=createaccount`

Additional parameters:
`username=SomeName`, `email=SomeEmail`, `realname=`, `mailpassword=1`

... and ensure you have a valid `createtoken` (use autofill in ApiSandbox)
and supply a `createreturnurl`.

After doing the above, the request should succeed with this patch applied
and should fail otherwise.

NOTE:
    - This has been tested locally to make sure previous scenarios
      where we used user types (temp account, anon account, etc)
      still works in the API environment without breaking like
      `clientlogin`, etc.
    - Reorganized `getRedirectUrl()` signature to pass in URL query
      params array instead of individual parameters for better
      management. The function param list was already growing too
      much.

Bug: T390751
Change-Id: I19f24b721f16f22437f8f946df771ec4da90d7e0
(cherry picked from commit 92c0008b2704ae91da7c5bf6b3e0e9bcbcbd2bc1)