Log a few bits of password entropy for login attempts
Log a very short hash of the password that's short enough
not to provide a meaningful amount of information even for weak
passwords, but long enough to be able to differentiate between
many failed login attempts with the same password (broken bots)
and iterating through different passwords (brute-force attack).

Salt with user data and add a secret and some extra length, so
not even a few bits of usable information about the password (or
the secret) is leaked. Given all that, no need to use a slow hash.

Also log standard request data.

Change-Id: I0ab031080e59577c9a93a0b618e37f6e838d1dd0