Re-enable AbuseFilterPreAuthenticationProvider on shared domain
We want to allow local abuse filter rules to prevail on the local
wiki regardless of whether a user succeeded in creating an account
from a different wiki.

A user can create an account on wiki A where a local abuse filter
rule allows them to create such an account with a given username
but disallowed on say wiki B. After the account is created on wiki
A, when the user tries to visit wiki B and click on the login link
which will trigger autocreation on the account created on wiki A
to be created on wiki B, the local abuse filter rules on wiki B
should prevent the account to be created because maybe it violates
some account naming abuse filter rules on that wiki.

This patch makes it so that abuse filter rules per wiki should always
run instead of ignored silently which will allow users to bypass
the rules and potentially cause harm.

Test plan
=========

0. Let SUL3 be enabled (some auth domain)
1. Have a local setup that has at least 2 wikis say A and B
2. Install AbuseFilter extension
3. Create an abuse filter rule on wiki B to block creation of
   accounts with a certain group of letters say "Test"
4. Visit wiki A and create an account say "TestAccount"
5. Logout and login with the newly created account on wiki A
6. Visit wiki B and click "Login" link on user menu
7. Autocreation of the "TestAccount" should fail because of
   the local abuse filter rule on that wiki to prevent account
   creation with "Test" in the account name.

Without this patch, step 7 will succeed and the user will be auto-
created and logged in which is not what we expected because the
rule is in place.

Bug: T393473
Change-Id: I4d0e85c7b832caf767b2e27443c83c9e9bb3b7c4