SECURITY: Do not count SUL3 autologin as reauthentication
In SUL3 mode, the local wiki has an authentication provider
which sends the user to the central login page, and if the user
logs in or is already logged in there, it considers that a
successful authentication and logs the user in locally.

But "logs in or is already logged in" is a big difference for
security reauthentication, where the entire point is to ensure
that the user proves their identity. Given how CentralAuth
sessions works, a successful session cookie theft will also
allow the attacker to pass the SUL3 equivalent of an autologin,
so those should not count for reauthentication.

We also need to make sure not to trigger a redirect loop
(user is sent to central login page, central login page sends
them back because they are already logged in, provider refuses
to set the reauth flag because there was no user interaction,
whatever page sent them to the local login page to reauthenticate
will just do so again). We do that by always forcing the central
login page to show in a reauthenticate flow. That also solves a
minor UX degradation in SUL3, as the presence of the 'force'
parameter triggers an explanatory message and that wasn't done on
the central login page previously.

There might be more elegant ways to handle the issue (since login
happens centrally, it would make sense to handle login recency
centrally as well, by storing the lastAuthTimestamp data in the
central session rather than the normal session) but those are too
complex to do in a security patch.

Bug: T389010
Depends-On: Ifce73837b25b0caad2d3d3cba000cceb0184c29d
Change-Id: I6937df05cd8dbb37d63b49147835467992843fe9