Fix WebAuthn sitenotice interacting with DismissableSiteNotice
We do not want the WebAuthn sitenotice to be dismissable.
Move the implementation to the 'SiteNoticeAfter' hook.

Note that the behavior depends on the order in which the
DismissableSiteNotice and CentralAuth extensions are loaded,
as they both implement handlers for the 'SiteNoticeAfter' hook now:

* If DismissableSiteNotice's 'SiteNoticeAfter' handler runs first,
  then CentralAuth's handler runs later and overwrites the dismissable
  elements' markup.

* If CentralAuth's 'SiteNoticeAfter' handler runs first, it returns
  `false` and aborts hook processing, so DismissableSiteNotice's
  handler is never called.

Both cases result in the desired effect.

Bug: T396424
Change-Id: Iedf3503d90775e4e9f07da2a53a8b587abd12aba