session: Remove `$expiry` variable, unused
This was introduced in 2016 with 0f38ef21774 (I27ccabdb8c365609)
but was never used even in that change set. Also, I think the
purpose was to pass it into clearCookies as the expire time in the
past (so that browswers will delete the relevant cookies).

But `clearCookie()` on the web response already does this with the
call to `WebResponse::setCookie()` with an expiry of:
`time() - 31_536_000` (1 year in the past).

So, the expiry in CentralAuthSessionProvider can be removed.

Change-Id: I73acc4ccb6e5f8f4218c55722507e3ade3e89aec