UnitTestsHandler: Fetch table names from `tables.json`
We don't want to create a maintainability burdun for developers
to always keep track of changes in `tables.json` and mirror it in
this constant. Let's read from the source of truth and if changes
occur there, the hook handler will just do the right thing without
any need to update the list of tables by hand.

This is inspired by I31a91b7d9db09103b6e5c58597 which I left a
comment there about why I like the approach used.

The constant was also public but not used by code outside of the
hook handler (should have been private already but well...).

Change-Id: If586a369be86a87e7de3cf6c50d36ba1448ae9c2