Show cached global edit counts in Special:CentralAuth and API
Everything else that checks users' global edit counts uses the cached
values (e.g. AbuseFilter variables, Wikipedia Library access control).
We should display them here as well, so that the behavior of those
tools is understandable. This also makes it possible for normal users
to identify discrepancies if the global edit count is out of sync for
whatever reason, without debugging the code.

The API change is also a performance improvement, as it no longer
needs to query every local wiki database to display the edit count.

Bug: T313900
Change-Id: I1bc1ad2059955d2a57ee373b66bd9a32fcdaa646