Enable using another wiki for SUL3 central login instead
This patch stops using 'loginwiki' as the configuration
source for SUL3's shared domain during central autologin.
We already have a concept in our production configuration
for setting `metawiki` as the central wiki. So, we can
introduce a new configuration and point that to `metawiki`,
then use that as the central wiki for SUL3 autologins.

See: https://codesearch.wmcloud.org/deployed/?q=%28%3D%7C%3D%3E%29+%27metawiki%27&files=&excludeFiles=&repos=operations%2Fmediawiki-config

Most things already use Meta-Wiki for administrative tasks
like global rename logging, account autocreation etc so it
seems safe to use this here(?).

Also, the loginwiki is configured different (in a non-standard
way) which might negatively affect authentication hence using
Meta-Wiki in this case which is a standard wiki and properly
configured.

NOTE: Per I89699d2979289eb (b160ff78363504cd9), we want to use
a wiki where we're sure that an account already exist for this
user and not autocreate on the local wiki. After normal account
creation, account autocreation is configured to happen on the
login wiki and meta-wiki[1]. So we could use that as the central
wiki.

Test plan
=========

0. Set `$wgCentralAuthCentralWiki = 'metawiki'`. `metawiki`
   should be a valid wiki in your wikifarm.
1. Visit the login page on a given SUL wiki
2. As a logged-out user, try to login in SUL3 mode (should be
   the default by now).
3. Monitor on the browser console that edge-login/autologin
   uses `metawiki` as the central wiki via URLs like:
   https://<shared-domain>/wiki/metawiki/...
4. You should login successfully with no failures.

NOTE: Account creation should also work fine with no failures.

[1] https://gerrit.wikimedia.org/g/operations/mediawiki-config/+/a2d2aaab9ace84280dd2f4c70a33bb69cd73850f/wmf-config/CommonSettings.php#2203

Bug: T402527
Depends-On: I96089d8139ec5adb951d954813a449733445df97
Change-Id: Ib5702b11b3ef642b6eda6e4c291c2fb670bc07f1