FixRenameUserLocalLogs: Skip rows where the performer is 'Global rename script'
These rows were created when the global performer's account did not
exist locally at the time of the rename. No need to do anything about
these rows, as they can't be affected by this bug, and we probably
shouldn't retroactively attribute them to the global performer if
their account was created locally in the meantime.

Bug: T398177
Change-Id: I820524f69cb188a165533fe52161131ff1ee1acc