SUL3: Remove passive domain in SUL3 authentication flow
During SUL2 -> SUL3 migration, we used the loginwiki sessions (SUL2)
as a fallback so that users don't get logged-out during rollout. But
the migration is since complete so we don't need to check this anymore
since SUL3 is the default active and it works nicely.

We also kept this in place so that in case of a rollback, SUL2 would
just work as before but this is no longer needed anymore as we're no
longer using SUL2 in production.

So this patch makes it such that we're either using SUL3 or SUL2
separately (as configured) and not using one as fallback to the other
as the migration is complete.

Test plan
=========

Before this patch:
- SUL3 authentication flow should fallback to SUL2 loginwiki during
  `/checkLoggedIn`

After this patch:
- SUL3 authentication flow should not fallback to SUL2 loginwiki
  during `/checkLoggedIn`
- There should be no fallback to SUL2 or a passive domain during
  SUL3. The authentication flow should be configured to use SUL2
  or SUL3 (separately and independently).
- Everything about login, signup etc should still work the same.

Bug: T403302
Change-Id: I7c156abe6fc20bf25562022bf9968bf9da517d45