session: Adapt consumer of SingleBackendSessionStore
Lookup which session backend implementation to use based on the
supplied config. Set anon cache type to false so that we use the
single backend store in this particular case.

The constructor of SingleBackendSessionStore is changing and this
avoids the backward compat hack with passing in null for stats
if we construct the object by hand.

Change-Id: I893c9d3ba0078e469a552a14ab1de57f8d127066